% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetScore.R
\name{GetScore}
\alias{GetScore}
\title{Obtain the score of each pathway through the gene expression matrix}
\usage{
GetScore(x, network, gamma = 0.5, select = 0.005, pathway)
}
\arguments{
\item{x}{A expression count matrix. The rows correspond to genes and
the columns correspond to cells.}

\item{network}{A adjacency matrix contation gene-gene interaction network.}

\item{gamma}{A number between 0 and 1 (default: 0.5).}

\item{select}{A threshold value, the absolute value of the score is less than
the value is regarded as an invalid value and will not participate
in the calculation}

\item{pathway}{pathway data}
}
\value{
The score of each path in the pathway data set
}
\description{
This function first randomly walks the gene expression matrix in
           the protein-protein interaction network, so that each gene in the
           PPI network gets a score. Afterwards, effective genes can be
           screened according to the set parameters. Finally, according to
           the pathway data in the pathway database,the score is matched with
           each gene after the score is obtained, and the score of each
           pathway is finally obtained.
}
